Move admin dashboard and menu into admin component
