+++ src/Monofony/Bundle/AdminBundle/Resources/config/services.yaml
@@ -3,7 +3,7 @@ services:
         autoconfigure: true
         autowire: false
 
-    Monofony\Bundle\AdminBundle\Dashboard\DashboardStatisticsProvider:
+    Monofony\Component\Admin\Dashboard\DashboardStatisticsProvider:
         arguments: [!tagged_iterator { tag: app.dashboard_statistic }]
 
-    Monofony\Bundle\AdminBundle\Dashboard\DashboardStatisticsProviderInterface: '@Monofony\Bundle\AdminBundle\Dashboard\DashboardStatisticsProvider'
+    Monofony\Component\Admin\Dashboard\DashboardStatisticsProviderInterface: '@Monofony\Component\Admin\Dashboard\DashboardStatisticsProvider'
