+++ src/Monofony/Component/Admin/composer.json
@@ -1,13 +1,13 @@
 {
-    "name": "monofony/core",
+    "name": "monofony/admin",
     "type": "library",
-    "description": "Monofony core.",
+    "description": "Monofony admin.",
     "license": "MIT",
     "require": {
         "php": "^7.2"
     },
     "autoload": {
-        "psr-4": { "Monofony\\Component\\Core\\": "" }
+        "psr-4": { "Monofony\\Component\\Admin\\": "" }
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
